SVNKIT-300: Merge branch 'url.parsing':
SVNKIT-300: Added more tests for other protocols parsing in order to make sure nothing is broken with the previous fix.

SVNKIT-300: URL parsing: fixes to ensure that "file:///" is a valid URL.

SVN considers file:/// as a valid URL corresponding to '/' path.
Until now SVNKit threw an exception on parsing of such an URL.
In practice this results into an impossibility to work even with file:///repo URL, because some operation try to get parent URL.